Random Number
Sets a random number between 1-20 and asks the user to guess which number we selected. Counts the number of guesses, and let's the user know when his guess is either too low or too high. When the user finally guesses the number, we display how many guesses it took him (as well as the correct number).